Fixed a few API issues + some other stuff

+++ api/register.php
@@ -7,16 +7,26 @@ $post = file_get_contents('php://input');
 $posted = json_decode($post, true);
 
 if (isset($posted)) {
-    $username = $posted['username'];
-    $password = $posted['password'];
-    $email = $posted['email'];
-
-    if (isset($posted['username']) && isset($posted['password']) && isset($posted['email'])) {
+    if (isset($posted['username']) && isset($posted['password']) && isset($posted['passwordConfirm']) && isset($posted['email'])) {
         $account = new Account();
+        $username = trim($posted['username']);
+        $password = trim($posted['password']);
+        $passwordConfirm = trim($posted['passwordConfirm']);
+        $email = $posted['email'];
+
+        if ($password != $passwordConfirm) {
+            echo json_encode(array('success' => false, 'message' => 'Passwords do not match.'));
+            exit;
+        }
+
+        if (strlen($password) < 8 || strlen($passwordConfirm) < 8 || strlen($username) < 4) {
+            echo json_encode(array('success' => false, 'message' => 'Username or password is too short.'));
+            exit;
+        }
         $res = $account->isAvailable($username, $email);
 
         if ($res) {
-            $response = $account->register($username, md5($password), $email);
+            $account->register($username, password_hash($password, PASSWORD_ARGON2ID), $email);
             echo json_encode(array('success' => true));
         } else {
             echo json_encode(array('success' => false, 'message' => 'Username or email already exists'));
